+++ packages/react/.storybook/main.js
@@ -75,7 +75,7 @@ module.exports = {
         alias: {
           ...config.resolve.alias,
           // we need an alias for hds-core to point webpack to the package as we can't use tilde (~) with rollup
-          './hds-core': require('path').resolve(__dirname, '../../../node_modules/hds-core'),
+          './hds-core': require('path').dirname(require.resolve('hds-core/package.json')),
         },
       },
     };